requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pymongo[srv]
dnspython
python-dotenv
python-multipart
pydantic
pydantic-settings
python-multipart
jinja2
